+++ plugin/src/adapter46/kotlin/org/jlleitschuh/gradle/ktlint/worker/BaselineLoader46.kt
@@ -1,9 +1,0 @@
-package org.jlleitschuh.gradle.ktlint.worker
-
-import com.pinterest.ktlint.core.internal.loadBaseline
-
-class BaselineLoader46 : BaselineLoader {
-    override fun loadBaselineRules(path: String): Map<String, List<SerializableLintError>> {
-        return loadBaseline(path).baselineRules?.mapValues { it.value.map { it.toSerializable() } } ?: emptyMap()
-    }
-}