+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintPlugin.kt
@@ -20,9 +20,6 @@ open class KtlintPlugin : Plugin<Project> {
 
     override fun apply(target: Project) {
         val holder = PluginHolder(target)
-        // Apply the idea plugin
-        target.plugins.apply(KtlintIdeaPlugin::class.java)
-
         holder.addKotlinScriptTasks()
         holder.addKtLintTasksToKotlinPlugin()
         holder.addGenerateBaselineTask()
@@ -42,7 +39,7 @@ open class KtlintPlugin : Plugin<Project> {
     private fun PluginHolder.applyKtlintMultiplatform(): (Plugin<in Any>) -> Unit = {
         val multiplatformExtension = target.extensions.getByType(KotlinMultiplatformExtension::class.java)
 
-        multiplatformExtension.sourceSets.all { sourceSet ->
+        multiplatformExtension.sourceSets.forEach { sourceSet ->
             val checkTask = createCheckTask(
                 this,
                 sourceSet.name,
@@ -73,7 +70,7 @@ open class KtlintPlugin : Plugin<Project> {
             addGenerateReportsTaskToProjectMetaFormatTask(generateReportsFormatTask)
         }
 
-        multiplatformExtension.targets.all { kotlinTarget ->
+        multiplatformExtension.targets.forEach { kotlinTarget ->
             if (kotlinTarget.platformType == KotlinPlatformType.androidJvm) {
                 applyKtLintToAndroid()
             }
@@ -81,16 +78,16 @@ open class KtlintPlugin : Plugin<Project> {
     }
 
     private fun PluginHolder.applyKtLint(): (Plugin<in Any>) -> Unit = {
-        target.extensions.configure<KotlinProjectExtension>("kotlin") { extension ->
-            extension.sourceSets.all { sourceSet ->
+        target.extensions.configure(KotlinProjectExtension::class.java) {
+            sourceSets.forEach { sourceSet ->
                 val kotlinSourceDirectories = sourceSet.kotlin.sourceDirectories
                 val checkTask = createCheckTask(
-                    this,
+                    this@applyKtLint,
                     sourceSet.name,
                     kotlinSourceDirectories
                 )
                 val generateReportsCheckTask = createGenerateReportsTask(
-                    this,
+                    this@applyKtLint,
                     checkTask,
                     GenerateReportsTask.LintType.CHECK,
                     sourceSet.name
@@ -100,12 +97,12 @@ open class KtlintPlugin : Plugin<Project> {
                 setCheckTaskDependsOnGenerateReportsTask(generateReportsCheckTask)
 
                 val formatTask = createFormatTask(
-                    this,
+                    this@applyKtLint,
                     sourceSet.name,
                     kotlinSourceDirectories
                 )
                 val generateReportsFormatTask = createGenerateReportsTask(
-                    this,
+                    this@applyKtLint,
                     formatTask,
                     GenerateReportsTask.LintType.FORMAT,
                     sourceSet.name