+++ samples/android-app/src/main/kotlin/org/jlleitschuh/gradle/ktlint/android/AlternativeSample.kt
@@ -1,5 +1,3 @@
 package org.jlleitschuh.gradle.ktlint.android
 
-data class AlternativeSample(
-    val one: Int = 1
-)
+data class AlternativeSample(val one: Int = 1)