+++ samples/kotlin-mpp-android/src/commonMain/kotlin/org/jlleitschuh/gradle/ktlint/sample/mppandroid/LocationProvider.kt
@@ -2,11 +2,13 @@ package org.jlleitschuh.gradle.ktlint.sample.mppandroid
 
 interface LocationProvider {
     fun startLocationFix()
+
     fun stopLocationFix()
+
     fun getCurrentLocation(): Location
 
     data class Location(
         val latitude: Long,
-        val longitude: Long
+        val longitude: Long,
     )
 }