+++ samples/kotlin-rulesets-creating/src/main/kotlin/org/jlleitschuh/gradle/ktlint/sample/kotlin/CustomRuleSetProvider.kt
@@ -5,7 +5,6 @@ import com.pinterest.ktlint.rule.engine.core.api.RuleProvider
 import com.pinterest.ktlint.rule.engine.core.api.RuleSetId
 
 class CustomRuleSetProvider : RuleSetProviderV3(RuleSetId("custom")) {
-
     override fun getRuleProviders(): Set<RuleProvider> {
         return setOf(RuleProvider { NoVarRule() })
     }